backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
distlib==0.3.9
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.14.0
filelock==3.16.1
html5lib==1.1
idna==3.10
lxml==5.3.0
mysqlclient==2.2.2
oscrypto==1.3.0
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
python-bidi==0.6.3
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
virtualenv==20.27.1
webencodings==0.5.1
xhtml2pdf==0.2.16
